Single book page
Route::resource('books', BookController::class)->only(['index', 'show']);

The code you provided is written in PHP and it defines a resourceful route for the "books" endpoint. This route is associated with the "BookController" class and allows only the "index" and "show" actions.

When you define a resourceful route in Laravel using the  Route::resource()  method, it automatically generates the necessary routes for typical CRUD operations on a resource. In this case, the "books" resource will have routes for listing all books (index action) and displaying a specific book (show action).

Here's a breakdown of the actions and corresponding routes that will be generated by the  Route::resource()  method for the "books" resource:

1. **Index Action**: This action retrieves and displays a list of all books.
   - Route: GET /books
   - Controller Method: BookController@index

2. **Show Action**: This action retrieves and displays a specific book.
   - Route: GET /books/{book}
   - Controller Method: BookController@show

Other CRUD actions such as create, store, edit, update, and delete can also be automatically generated by the  Route::resource()  method if you include them in the array passed as the second argument. However, in your code, only the "index" and "show" actions are specified using the  only()  method.

By using resourceful routing, you can easily define and manage routes for CRUD operations on a resource in Laravel.

How to remove the ? betwn the book and the book ID for the single page

In Laravel, the default behavior for generating URLs with route parameters is to use the "/" symbol to separate the segments of the URL. However, in some cases, you may want to use a query string parameter instead, especially when dealing with optional or additional parameters.

To generate a URL with the "/" symbol instead of the "?" symbol, you can modify the route definition or the way the URL is generated. Here's an example:

1. Modify the route definition:
   Instead of using the  Route::resource()  method, you can define a custom route that explicitly specifies the URL structure using the "/" symbol. For example:
php
Route::get('books/{book}', [BookController::class, 'show'])->name('books.show');
2. Generate the URL with the "/" symbol:
   Instead of using the  route()  function with the default behavior, you can use the  url()  function to generate the URL with the "/" symbol. For example:
php
   <a href="{{ url('books', $book) }}" class="book-title">{{ $book->title }}</a>

By making these modifications, you can generate a URL with the "/" symbol instead of the "?" symbol for separating the ID parameter.